+++ packages/cms/lib/modules/resource-overview-widgets/lib/fields.js
@@ -604,7 +604,6 @@ module.exports = [
     name: 'siteId',
     label: 'Site ID',
   },
-
 ].concat(
   ideaStates.map((state) => {
     return {